update: record post

See merge request s06-final/S06P31A301!32

+++ server/src/main/java/com/darly/db/entity/record/Record.java
@@ -30,7 +30,7 @@ public class Record {
     private Long recordDate;
     private Float recordDistance;
     private Long recordTime;
-    private Float recordPace;
+    private Integer recordPace;
     private Float recordSpeed;
     private Integer recordHeart;
     private Float recordStartLatitude;
